refactor: DB 접속 계층 리팩토링
- persistence -> infrastructure 이름 변경
- existsByLoginId 메서드 추가
- fetch 메서드 추가
