+++ src/main/java/com/flab/just_10_minutes/User/infrastructure/UserDao.java
@@ -1,12 +1,15 @@
-package com.flab.just_10_minutes.User.persistence;
+package com.flab.just_10_minutes.User.infrastructure;
 
 import com.flab.just_10_minutes.User.domain.User;
 import com.flab.just_10_minutes.Util.Exception.Database.DuplicatedKeyException;
 import com.flab.just_10_minutes.Util.Exception.Database.InternalException;
+import com.flab.just_10_minutes.Util.Exception.Database.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
+
+import static com.flab.just_10_minutes.Util.contants.ResponseMessage.*;
 
 @Repository
 @RequiredArgsConstructor
@@ -23,10 +26,21 @@ public class UserDao {
             int insertCount = userMapper.save(user);
 
             if (insertCount != 1) {
-                throw new InternalException("Fail to Insert");
+                throw new InternalException(FAIL_TO_INSERT);
             }
         } catch (DuplicateKeyException e) {
-            throw new DuplicatedKeyException("Duplicate User loginId for Registration");
+            throw new DuplicatedKeyException(DUPLICATED_KEY_LOGIN_ID);
         }
+    }
+
+    public boolean existsByLoginId(final String loginId) {
+        return userMapper.existsByLoginId(loginId);
+    }
+
+    public User fetch(final String loginId) {
+        if (!userMapper.existsByLoginId(loginId)) {
+            throw new NotFoundException(NOT_EXIST_USER);
+        }
+        return userMapper.findByLoginId(loginId);
     }
 }
